requirements.txt
asgiref==3.4.1
Django==3.2.5
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
